refactor(multisig-config): move to strata-crypto
