+++ crates/asm/subprotocols/upgrade/src/state.rs
@@ -1,7 +1,6 @@
 use borsh::{BorshDeserialize, BorshSerialize};
-use strata_asm_proto_upgrade_txs::{
-    actions::UpdateId, crypto::multisig_config::MultisigConfigUpdate,
-};
+use strata_asm_proto_upgrade_txs::actions::UpdateId;
+use strata_crypto::multisig::config::MultisigConfigUpdate;
 use strata_primitives::roles::Role;
 
 use crate::{
