+++ crates/asm/txs/upgrade/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2024"
 
 [dependencies]
 strata-asm-common.workspace = true
+strata-crypto.workspace = true
 strata-primitives.workspace = true
 
 arbitrary.workspace = true
