+++ crates/asm/txs/upgrade/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod actions;
 pub mod constants;
-pub mod crypto;
 pub mod error;
 pub mod parse;
