+++ crates/crypto/src/lib.rs
@@ -3,3 +3,4 @@
 // FIXME this stub had to be moved to make a refactor work
 pub use strata_primitives::crypto::*;
 pub mod groth16_verifier;
+pub mod multisig;
