+++ crates/crypto/src/multisig/config.rs
@@ -1,7 +1,7 @@
 use arbitrary::Arbitrary;
 use borsh::{BorshDeserialize, BorshSerialize};
 
-use crate::{crypto::PubKey, error::MultisigConfigError};
+use crate::multisig::{errors::MultisigConfigError, PubKey};
 
 /// Configuration for a multisignature authority:
 /// who can sign (`keys`) and how many of them must sign (`threshold`).
