+++ crates/crypto/src/multisig/mod.rs
@@ -1,11 +1,12 @@
 // TODO: This needs to be in a different crate. Maybe strata-crypto
+pub mod config;
+pub mod errors;
 pub mod msg;
-pub mod multisig_config;
 pub mod vote;
 
 use strata_primitives::buf::{Buf32, Buf64};
 
-use crate::error::VoteValidationError;
+use crate::multisig::{errors::VoteValidationError, msg::MultisigPayload};
 
 pub type PubKey = Buf32;
 pub type Signature = Buf64;
@@ -16,6 +17,6 @@ pub fn aggregate_pubkeys(_keys: &[PubKey]) -> Result<PubKey, VoteValidationError
 }
 
 // FIXME: handle
-pub fn verify_sig(_pk: &PubKey, _msg_hash: &Buf32, _sig: &Signature) -> bool {
+pub fn verify_sig(_pk: &PubKey, _payload: &MultisigPayload, _sig: &Signature) -> bool {
     true
 }
